requirements.txt
requests>=2.31.0
pyyaml>=6.0.1
icalendar>=5.0.11
python-dateutil>=2.8.2